Added Optimized Article Card2 + overview in profile
