+++ client/app/articles/page.tsx
@@ -1,6 +1,6 @@
 import Article from "@/models/article";
 import Tag from "@/models/tag";
-import ArticleCard from "../components/testCard";
+import ArticleCard from "../components/ArticleCard2";
 export default async function Articles() {
   // const res = await fetch("http://localhost:8080/article/published", {
   //   method: "GET",
@@ -57,7 +57,7 @@ export default async function Articles() {
         <p>اخر المقالات</p>
       </div>
       <div className="container my-12 mx-auto px-4 md:px-12">
-        <div className="flex flex-wrap justify-center -mx-1 lg:-mx-4 text-content">
+        <div className="flex flex-wrap justify-center gap-10 md:gap-4 mx-1 lg:-mx-4 text-content">
         {articles.map((article) => {
           return (
                 <ArticleCard article={article} key={article.article_id} />
